Create spatial sf object from start and end point of line transects with eastings and northings

Create spatial sf object from start and end point of line transects with eastings and northings

Usage

fcn_line_transect_sf(
  lineTransect,
  colnames = list(Start_Eastings = "Start_Eastings", Start_Northings = "Start_Northings",
    End_Eastings = "End_Eastings", End_Northings = "End_Northings"),
  start_point = F
)

Arguments

lineTransect

A dataframe with columns of Start_Eastings, Start_Northings, End_Eastings, and End_Northings

colnames

A list with items Start_Eastings, Start_Northings, End_Eastings, and End_Northings, which are strings corresponding to the column name

start_point

A boolean indicating whether the output sf should be just the start points, or the whole table